requirements.txt
Django==3.1.3
django-extensions==3.0.9
django-filter==2.4.0
django-health-check==3.14.3
djangorestframework==3.12.1
djangorestframework-jwt==1.11.0
drf-yasg==1.20.0
gunicorn==20.0.4
Markdown==3.3.3
psutil==5.7.3
pytest==6.1.2
pytest-django==4.1.0
